requirements.txt
torch==2.6.0
torchvision==0.21.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
basicsr==1.3.5
beautifulsoup4==4.13.4
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.2
camel-ai==0.2.70
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clint==0.5.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
coloredlogs==15.0.1
config
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
dashscope==1.23.8
datasets==4.0.0
diffusers @ git+https://github.com/huggingface/diffusers.git@3c8b67b3711b668a6e7867e08b54280e51454eb5
dill==0.3.8
distro==1.9.0
docstring-parser==0.15
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
fairscale==0.4.13
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fire==0.4.0
flatbuffers==25.2.10
# Editable install with no version control (flux==0.0.post58+g1371b2b)
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gradio==5.13.2
gradio_client==1.6.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hpsv2==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.10
image-reward==1.5
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.1.0
invisible-watermark==0.2.0
iopath==0.1.10
insightface
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.7.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mcp==1.11.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==1.46.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
onnx==1.14.0
onnxruntime==1.19.0
open_clip_torch==2.32.0
openai==1.95.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.6.8
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.4.1
starlette==0.47.1
streamlit==1.46.1
streamlit-drawable-canvas==0.9.3
streamlit-keyup==0.3.0
stringzilla==3.12.5
sympy==1.13.1
t2v_metrics @ git+https://github.com/linzhiqiu/t2v_metrics.git
tb-nightly==2.20.0a20250714
tenacity==9.1.2
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
transformers
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.167
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webdataset==1.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0